fix hook
