+++ packages/client/src/feature/session/register/hooks/useRegister.tsx
@@ -1,10 +1,10 @@
-import { FormEvent } from 'react'
+import { useState } from 'react'
 
-import { useFormData } from '@/shared/components/core/FormData/hooks/useFormData'
-import { FormDataPayload } from '@/shared/components/core/FormData/types'
+import { FormDataError, FormDataPayload } from '@/shared/types'
+import { validateField } from '@/shared/components/core/FormData/model/validateField'
 
 export const useRegister = () => {
-  const initialState = {
+  const [formData, setFormData] = useState<FormDataPayload>({
     first_name: '',
     second_name: '',
     login: '',
@@ -12,17 +12,36 @@ export const useRegister = () => {
     phone: '',
     password: '',
     password_repeat: ''
+  })
+
+  const [errors, setErrors] = useState<FormDataError>({
+    first_name: '',
+    second_name: '',
+    login: '',
+    email: '',
+    phone: '',
+    password: '',
+    password_repeat: ''
+  })
+
+  const isError = Object.values(errors).some(error => error !== '')
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target
+    setFormData({ ...formData, [name]: value })
+    setErrors({ ...errors, [name]: validateField(name, value, formData.password) })
   }
 
-  const { inputProps, handleSubmit, isError } = useFormData(initialState)
-
-  const submitRegister = (data: FormDataPayload) => {
-    console.log(`Отправка формы... \n ${JSON.stringify(data, null, 2)}`)
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault()
+    console.log(`Отправка формы... \n ${JSON.stringify(formData, null, 2)}`)
   }
+
+  const inputProps = { formData, handleChange, errors }
 
   return {
     inputProps,
-    handleSubmit: (e: FormEvent) => handleSubmit(e, submitRegister),
+    handleSubmit,
     isError
   }
 }
